move setAPI to class Util
