+++ app/src/main/java/vn/asiantech/travelmate/utils/SetAPIUtil.kt
@@ -1,0 +1,25 @@
+package vn.asiantech.travelmate.utils
+
+import com.google.gson.GsonBuilder
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import vn.asiantech.travelmate.detailactivity.SOService
+
+class SetAPIUtil {
+
+    fun setUpApi(url: String): SOService {
+        val logging = HttpLoggingInterceptor()
+        logging.level = HttpLoggingInterceptor.Level.BODY
+        val httpClient = OkHttpClient.Builder()
+        httpClient.addInterceptor(logging)
+        val gson = GsonBuilder().setLenient().create()
+        val getImagesRetrofit = Retrofit.Builder()
+            .baseUrl(url)
+            .addConverterFactory(GsonConverterFactory.create(gson))
+            .client(httpClient.build())
+            .build()
+        return getImagesRetrofit.create<SOService>(SOService::class.java)
+    }
+}
